css/css-regions/parsing/flow-into-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Regions Module Level 1: getComputedValue().flowInto</title>
<link rel="help" href="https://drafts.csswg.org/css-regions/#propdef-flow-into">
<meta name="assert" content="flow-into computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("flow-into", "none");

test_computed_value("flow-into", "COLUMN");
test_computed_value("flow-into", "COLUMN element");
test_computed_value("flow-into", "COLUMN content");
</script>
</body>
</html>
